requirements.txt
customtkinter>=5.0.0
yt-dlp
requests
python-dotenv